requirements.txt
PyMySQL==1.1.0
PyYAML==6.0
allure-pytest==2.13.2
allure-pytest-commons==2.13.2
deepdiff==6.3.1
jsonpath==0.82
jsonpath-rw==1.4.0
pytest==7.3.1
requests==2.31.0
requests-toolbelt==1.0.0
simplejson==3.19.1
python-dateutil==2.8.2







pytest==6.0.2
pytest-html==3.0.0
pytest-xdist==2.1.0
pytest-rerunfailures==9.1.1
allure-pytest==2.8.18
requests==2.24.0
requests-toolbelt==0.9.1
simplejson==3.17.2
ruamel.yaml==0.16.12
PyYAML==5.3.1
configparser==5.0.0
python-dateutil==2.8.0
#elasticsearch==7.13.4
#phoenixdb==1.0.1
#pymysql==1.0.2
#pysolr==3.8.1
#redis==3.5.3